Refactor: Split pages.scss into separate files for each page for better organization and readability
