+++ src/App.jsx
@@ -1,16 +1,15 @@
 import React from "react";
 import "./App.css";
 import { Routes, Route } from "react-router-dom";
-import HomePage from "./pages/HomePage";
-import DashboardPage from "./pages/DashboardPage";
-import AssetPage from "./pages/AssetPage";
-import ReportPage from "./pages/ReportPage";
+import HomePage from "./pages/HomePage/HomePage.jsx";
+import AssetPage from "./pages/AssetPage/AssetPage.jsx";
+import ReportPage from "./pages/ReportPage/ReportPage.jsx";
 import Layout from "./shared/Layout";
 import Login from "./components/Login/Login.jsx";
 import PrivateRoute from "./core/guard/privateRoute.jsx";
-import AssetDetailPage from "./pages/AssetDetailPage.jsx";
-import FloormapDetail from "./pages/FloormapDetail.jsx";
-import ZoneEditing from "./pages/ZoneEditingPage.jsx";
+import AssetDetailPage from "./pages/AssetDetailPage/AssetDetailPage.jsx";
+import FloormapDetail from "./pages/FloormapDetailPage/FloormapDetailPage.jsx";
+import ZoneEditing from "./pages/ZoneEditingPage/ZoneEditingPage.jsx";
 
 function App() {
   return (
@@ -19,10 +18,6 @@ function App() {
         <Routes>
           <Route path="/" element={<HomePage />} />
           <Route path="/login" element={<Login />} />
-          <Route
-            path="/dashboard"
-            element={<PrivateRoute element={DashboardPage} />}
-          />
           <Route
             path="/floormap/:floormapId"
             element={<PrivateRoute element={FloormapDetail} />}
