+++ src/pages/AssetDetailPage/_assetDetailPage.scss
@@ -1,0 +1,1 @@
+@use "../../styles/variables";
