+++ src/pages/AssetPage/AssetPage.jsx
@@ -1,12 +1,12 @@
 import React, { useEffect, useState } from "react";
-import { AssetService } from "../services/assetService.js";
-import AssetTable from "../components/AssetTable/AssetTable.jsx";
-import AddAssetForm from "../components/AddAssetForm/AddAssetForm.jsx";
-import Modal from "../components/Modal/Modal.jsx";
-import Pagination from "../components/Pagination/Pagination.jsx";
-import "./_pages.scss";
-import { AssetPropType } from "../core/types/assetPropType.js";
-import { FloorMapService } from "../services/floormapService.js";
+import { AssetService } from "../../services/assetService.js";
+import AssetTable from "../../components/AssetTable/AssetTable.jsx";
+import AddAssetForm from "../../components/AddAssetForm/AddAssetForm.jsx";
+import Modal from "../../components/Modal/Modal.jsx";
+import Pagination from "../../components/Pagination/Pagination.jsx";
+import "./_assetPage.scss";
+import { AssetPropType } from "../../core/types/assetPropType.js";
+import { FloorMapService } from "../../services/floormapService.js";
 
 function AssetPage() {
   const [assets, setAssets] = useState([]);
@@ -64,13 +64,13 @@ function AssetPage() {
 
   const handleAddingAsset = async (newAsset) => {
     try {
-      console.log("New asset:", newAsset)
+      console.log("New asset:", newAsset);
       const addedAsset = await AssetService.addAsset(newAsset);
       setAssets([...assets, addedAsset]);
     } catch (error) {
       console.error("Error adding asset:", error.message);
     }
-  }
+  };
 
   return (
     <div className="asset-page">
@@ -80,9 +80,7 @@ function AssetPage() {
         </div>
         <div className="asset-container__content">
           <Modal buttonText="Add asset" title="Add new asset">
-            <AddAssetForm
-                onAddAsset={handleAddingAsset}
-            />
+            <AddAssetForm onAddAsset={handleAddingAsset} />
           </Modal>
         </div>
       </div>
