+++ src/pages/AssetPage/_assetPage.scss
@@ -1,4 +1,4 @@
-@use "../styles/variables";
+@use "../../styles/variables";
 
 .asset-container {
   display: flex;
