--- src/pages/DashboardPage.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-function DashboardPage() {
-  return <div>Dashboard</div>;
-}
-
-export default DashboardPage;
