+++ src/pages/HomePage/_homePage.scss
@@ -1,0 +1,1 @@
+@use "../../styles/variables";
