+++ src/pages/ReportPage/_reportPage.scss
@@ -1,0 +1,1 @@
+@use "../../styles/variables";
